fix: add search queries to ListSessions and remodel integration tests
